+++ packages/ui/src/selectors/DataSourceSelector/utils/index.ts
@@ -1,10 +1,10 @@
 import { compact, find, isArray, isEmpty, map } from "lodash";
-import { DATA_SOURCES } from "../constants/index.js";
+import { DATA_SOURCES } from "../constants";
 import CustomFunctions from "../models/customFunctions.js";
 import DataSets from "../models/dataSets.js";
 import DataSourceModel from "../models/dataSource.js";
 import NativeDataSource from "../models/nativeDataSource.js";
-import { DataSourceType } from "../types/index.js";
+import { DataSourceType } from "../types";
 import { SqlViews } from "../models/sqlViews.js";
 
 export function getDataSourcesList(
@@ -32,7 +32,7 @@ export function getDataSourcesList(
 					}
 				}
 			}),
-		);
+		) as DataSourceModel[];
 	}
 	return [];
 }