+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hisptz/dhis2-ui
 
+## 2.0.22
+
+### Patch Changes
+
+- Data source selector fix
+
 ## 2.0.21
 
 ### Patch Changes